restructure and add multilabl classification
