+++ README.md
@@ -35,20 +35,20 @@ lf_associations = []
 for text_id, text in enumerate(texts):
     label = contains_keywords(text)
     if label is not None:
-        association = weak_nlp.ClassificationAssociation(text_id + 1, label)
+        association = weak_nlp.SingleClassificationAssociation(text_id + 1, label)
         lf_associations.append(association)
 
 lf_vector = weak_nlp.SourceVector(contains_keywords.__name__, False, lf_associations)
 
 ground_truths = [
-    weak_nlp.ClassificationAssociation(1, "clickbait"),
-    weak_nlp.ClassificationAssociation(2, "regular"),
-    weak_nlp.ClassificationAssociation(3, "regular")
+    weak_nlp.SingleClassificationAssociation(1, "clickbait"),
+    weak_nlp.SingleClassificationAssociation(2, "regular"),
+    weak_nlp.SingleClassificationAssociation(3, "regular")
 ]
 
 gt_vector = weak_nlp.SourceVector("ground_truths", True, ground_truths)
 
-cnlm = weak_nlp.CNLM([gt_vector, lf_vector])
+cnlm = weak_nlp.SingleLabelClassificationNLM([gt_vector, lf_vector])
 ```
 
 Whereas for extraction tasks, your code snippet could look as follows:
@@ -66,19 +66,19 @@ lf_associations = []
 for text_id, text in enumerate(texts):
     for triplet in match_keywords(text):
         label, from_idx, to_idx = triplet
-        association = weak_nlp.ExtractionAssociation(text_id + 1, label, from_idx, to_idx)
+        association = weak_nlp.ExclusiveEntityExtractionAssociation(text_id + 1, label, from_idx, to_idx)
         lf_associations.append(association)
 
 lf_vector = weak_nlp.SourceVector(match_keywords.__name__, False, lf_associations)
 
 ground_truths = [
-    weak_nlp.ExtractionAssociation(1, "person", 1, 2),
-    weak_nlp.ExtractionAssociation(2, "person", 4, 5),
+    weak_nlp.ExclusiveEntityExtractionAssociation(1, "person", 1, 2),
+    weak_nlp.ExclusiveEntityExtractionAssociation(2, "person", 4, 5),
 ]
 
 gt_vector = weak_nlp.SourceVector("ground_truths", True, ground_truths)
 
-enlm = weak_nlp.ENLM([gt_vector, lf_vector])
+enlm = weak_nlp.ExclusiveEntityExtractionNLM([gt_vector, lf_vector])
 ```
 
 ## Roadmap
