+++ weak_nlp/__init__.py
@@ -1,3 +1,13 @@
-from weak_nlp.classification import ClassificationAssociation, CNLM
-from weak_nlp.extraction import ExtractionAssociation, ENLM
+from weak_nlp.classification.single_label import (
+    SingleClassificationAssociation,
+    SingleLabelClassificationNLM,
+)
+from weak_nlp.classification.multi_label import (
+    MultilabelMulticlassAssociation,
+    MultiLabelMultiClassNLM,
+)
+from weak_nlp.extraction.exclusive import (
+    ExclusiveEntityExtractionAssociation,
+    ExclusiveEntityExtractionNLM,
+)
 from weak_nlp.base import SourceVector
